Update helm charts path in deployment configuration
Changed helm charts path from 'helmfiles' to 'helm-charts' in deployment.yml. This modification is necessary as the path to the helm charts in the deployment configuration was incorrect. This prevented the successful deployment of the backend due to the helm upgrade command not being able to locate the proper charts. The issue has now been resolved with this update.