Add metadata labels and update helper template in helm charts
Added missing metadata labels in deployment.yaml to improve Kubernetes resource identification. Additionally, removed the "aznopoly-server.serviceAccountName" block in _helpers.tpl as it is no longer necessary. The service account name is nested in our new deployments configuration and helm charts. This update ensures the consistency and correctness of the helm charts and eases the helm upgrade process by pointing to the correct chart location.