Refactor server logics and introduces RoomManager, UUID based user system.
Refactored server logic in the source index file to now utilize a brand new RoomManager class and UUID based identification system, rather than relying on a username header. RoomManager class allows the system to easily manage joins and leaves of a room, as well optimal handling and removal of empty rooms. UUID based user identification was chosen over the previous username system to ensure uniqueness of each user and to avoid potential collisions. This will greatly improve the consistency and performance of the system.