Change return type of getRoom method from string to object
Modified return type of getRoom method in room_manager.ts from string to object. Instead of converting the returned information into a JSON string it now returns the information as is in an object. This simplifies the data processing on the client’s end by removing the necessity to parse the string back to an object. This change also ensures type consistency which enhances code maintainability and reduces potential type-related issues.